fix: overhaul how TypeScript variants are applied to ensure different frontends can be used together
This ensures that the Stimulus variant will generate a valid TypeScript-based project when the React
 and Bootstrap variants are _not_ enabled, and also resolves a conflict between the React and
 Bootstrap variants in `application.ts`.
