+++ ci/configs/basic-typescript.yml
@@ -1,0 +1,12 @@
+---
+staging_hostname: "staging.example.com"
+production_hostname: "www.example.com"
+git_repo_url: ""
+use_typescript: true
+apply_variant_github_actions_ci: false
+apply_variant_react: false
+apply_variant_devise: false
+apply_variant_sidekiq: false
+apply_variant_bootstrap: false
+apply_variant_deploy_with_capistrano: false
+apply_variant_deploy_with_ackama_ec2_capistrano: false
