+++ variants/frontend-base-typescript/.eslintrc.js
