+++ variants/frontend-base-typescript/tsconfig.json
