+++ variants/frontend-base-typescript/types.d.ts
