--- variants/frontend-react-typescript/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES6",
-    "module": "ES6",
-    "moduleResolution": "Node",
-    "lib": ["ES6", "ES2020", "DOM"],
-    "jsx": "react",
-    "noEmit": true,
-    "noEmitOnError": true,
-    "allowJs": true,
-    "checkJs": true,
-    "strict": true,
-    "useDefineForClassFields": true,
-    "forceConsistentCasingInFileNames": true,
-    "noImplicitReturns": true,
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true
-  },
-  "exclude": ["coverage", "node_modules", "vendor", "public"],
-  "files": [".eslintrc.js"],
-  "include": ["**/*.js", "**/*.ts", "**/*.tsx"]
-}
